backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
Pillow==9.3.0
djoser==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
django-filter==21.1
python-dotenv==0.21.1
urllib3==1.26.15
django-import-export
drf-extra-fields==3.4.0
pytz==2021.1
sqlparse==0.4.1
